+++ app/javascript/vue/tasks/metadata/vocabulary/project_vocabulary/components/TableWords.vue
@@ -2,8 +2,8 @@
   <table class="table-striped">
     <thead>
       <tr>
-        <th>Term</th>
-        <th>Count</th>
+        <th @click="sortList(0)">Term</th>
+        <th @click="sortList(1)">Count</th>
       </tr>
     </thead>
     <tbody>
@@ -20,12 +20,20 @@
 </template>
 
 <script setup>
-defineProps({
-  list: {
-    type: Array,
-    default: () => []
-  }
+import { ref } from 'vue'
+import { sortArray } from '@/helpers'
+
+const list = defineModel({
+  type: Array,
+  default: () => []
 })
 
-const emit = defineEmits(['select'])
+const emit = defineEmits(['select', 'sort'])
+
+const asc = ref(true)
+
+function sortList(position) {
+  list.value = sortArray(list.value, position, asc.value)
+  asc.value = !asc.value
+}
 </script>